+++ tests/test_queries/test_nested_reverse_relations.py
@@ -56,12 +56,8 @@ def create_test_database():  # pragma: no cover
     metadata.drop_all(engine)
 
 
-@pytest.mark.skipif(
-    database._backend._dialect.name == "sqlite",
-    reason="wait for fix for sqlite in encode/databases",
-)
 @pytest.mark.asyncio
-async def test_double_nested_reverse_relation():  # pragma: no cover
+async def test_double_nested_reverse_relation():
     async with database:
         data_source = await DataSource(name="local").save()
         test_tables = [